add path to complx-test call; fix type on signum

+++ wybelibs/wybe/cbits.c
@@ -93,7 +93,7 @@ int64_t isqrt(int64_t x) {
     return (int64_t)s;
 }
 
-int64_t signum(double x) {
+int64_t signum(int64_t x) {
     return (x > 0) - (x < 0); 
 }
